manage trip-check booking handled-

+++ Rahhal-Backend/src/main/java/com/rahhal/service/TripServiceImpl.java
@@ -3,7 +3,6 @@ package com.rahhal.service;
 import com.rahhal.dto.TripDto;
 import com.rahhal.entity.Trip;
 import com.rahhal.entity.User;
-import com.rahhal.enums.UserRole;
 import com.rahhal.exception.EntityNotFoundException;
 import com.rahhal.repository.TripRepository;
 import com.rahhal.repository.UserRepository;
@@ -13,7 +12,7 @@ import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.stereotype.Service;
 
 @Service
-public class TripServiceImpl implements TripService{
+public class TripServiceImpl implements TripService {
     private final TripRepository tripRepository;
     private final UserRepository userRepository;
 
@@ -67,7 +66,10 @@ public class TripServiceImpl implements TripService{
             throw new RuntimeException("Unauthorized to update this trip");
         }
 
-        // Don't forget check booking
+        if(trip.isBooked())
+        {
+            throw new IllegalStateException("Can't update the trip is booked");
+        }
 
         trip.setTitle(tripDto.getTitle());
         trip.setDescription(tripDto.getDescription());
@@ -92,7 +94,10 @@ public class TripServiceImpl implements TripService{
             throw new RuntimeException("Unauthorized to update this trip");
         }
 
-        // Don't forget check booking
+        if(trip.isBooked())
+        {
+            throw new IllegalStateException("Can't delete the trip is booked");
+        }
 
         tripRepository.delete(trip);
     }
